requirements.txt
# Python dependencies for Smart Discord Music Bot

# Discord
discord.py[voice]>=2.3.0

# YouTube
yt-dlp>=2024.1.0
ytmusicapi>=1.3.0

# Spotify
spotipy>=2.23.0

# Database
aiosqlite>=0.19.0

# Web Dashboard
aiohttp>=3.9.0

# Configuration
python-dotenv>=1.0.0

# Encryption (for future auth features if needed)
cryptography>=41.0.0

# Audio processing (PyNaCl is required for voice)
PyNaCl>=1.5.0
psutil>=5.9.0
